Move binary impls into kernels/binary.rs
