+++ candle-core/src/metal_backend/mod.rs
@@ -1832,7 +1832,7 @@ impl MetalStorage {
         let lhs = buffer_o(&self.buffer, lhs_l, self.dtype);
         let rhs = buffer_o(&rhs.buffer, rhs_l, rhs.dtype);
         let (buffer, dtype) = if lhs_l.is_contiguous() && rhs_l.is_contiguous() && &op[..1] != "b" {
-            use candle_metal_kernels::binary::contiguous;
+            use candle_metal_kernels::kernels::binary::contiguous;
 
             let (kernel_name, dtype) = match (op, self.dtype) {
                 ("add", DType::F32) => (contiguous::add::FLOAT, self.dtype),
@@ -1919,7 +1919,7 @@ impl MetalStorage {
             .map_err(MetalError::from)?;
             (buffer, dtype)
         } else {
-            use candle_metal_kernels::binary::strided;
+            use candle_metal_kernels::kernels::binary::strided;
 
             let (kernel_name, dtype) = match (op, self.dtype) {
                 ("badd", DType::F32) => (strided::add::FLOAT, self.dtype),
